Ported matrixmul demo to 64-bit

+++ demos/matrixmul/src/application/matrix_functions.c
@@ -1,41 +1,42 @@
 /* matrix_functions.c */
 
-#include "matrix_functions.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <arpa/inet.h>
 
 #include <assert.h>
 
-void add_matrixes(int* target, int *source_a, int *source_b, int matrix_size) {
+#include "matrix_functions.h"
+
+void add_matrixes(int64_t* target, int64_t *source_a, int64_t *source_b, int matrix_size) {
 	int i;
 	for (i=0; i<(matrix_size*matrix_size); ++i) {
 		target[i] = source_a[i] + source_b[i];
 	}
 }
 
-void sub_matrixes(int* target, int *source_a, int *source_b, int matrix_size) {
+void sub_matrixes(int64_t* target, int64_t *source_a, int64_t *source_b, int matrix_size) {
 	int i;
 	for (i=0; i<(matrix_size*matrix_size); ++i) {
 		target[i] = source_a[i] - source_b[i];
 	}
 }
 
-void copy_matrix(int* target, int *source, int matrix_size) {
+void copy_matrix(int64_t* target, int64_t *source, int matrix_size) {
 	int i;
 	for (i=0; i<(matrix_size*matrix_size); ++i) {
 		target[i] = source[i];
 	}
 }
 
-int read_matrix(int* target, char *source_file, int matrix_size){
+int read_matrix(int64_t* target, char *source_file, int matrix_size){
 	// WARNING: no path string sanitizing! Use with caution!
 	FILE* mf = fopen(source_file, "r");
 
 	assert(target != NULL);
 	assert(source_file != NULL);
 
-	size_t read_items = fread(target, sizeof(int), matrix_size*matrix_size, mf);
+	size_t read_items = fread(target, sizeof(int64_t), matrix_size*matrix_size, mf);
 	fclose(mf);
 
 	// Convert from network byte order to host byte order (endianness)
@@ -47,11 +48,11 @@ int read_matrix(int* target, char *source_file, int matrix_size){
 	return (read_items == (matrix_size*matrix_size) );
 }
 
-int write_matrix(int* source, char *target_file, int matrix_size){
+int write_matrix(int64_t* source, char *target_file, int matrix_size){
 	// WARNING: no path string sanitizing! Use with caution!
 
 	// Convert field from host byte order to network byte order (endianness)
-	int* source2 = malloc(matrix_size*matrix_size*sizeof(int));
+	int64_t* source2 = malloc(matrix_size*matrix_size*sizeof(int64_t));
 	unsigned long i;
 	for(i = 0; i< matrix_size*matrix_size; i++){
 		source2[i] = htonl(source[i]);
@@ -59,7 +60,7 @@ int write_matrix(int* source, char *target_file, int matrix_size){
 
 	// Write field to file
 	FILE* mf = fopen(target_file, "w");
-	size_t written_items = fwrite(source2, sizeof(int), matrix_size*matrix_size, mf);
+	size_t written_items = fwrite(source2, sizeof(int64_t), matrix_size*matrix_size, mf);
 	fclose(mf);
 
 	free (source2);
